Add SamPredictorWrapper class and utility functions
This commit adds a new class called SamPredictorWrapper, which is a wrapper around the SamPredictor class from the segment_anything module. The SamPredictorWrapper class provides a convenient interface for predicting masks on images using the SamPredictor model.

In addition to the class, this commit also introduces two utility functions: save_mask_as_image and save_mask. These functions are used to save the predicted masks as images or .npy files, respectively.

The commit also includes an example usage of the SamPredictorWrapper class in the main.py file.